rf: move `polish_validation_results()` and supporting defs
From `cli/models.py` to `models.py`
